Refactor util.js into modules
