+++ templates/src/js/utils/enableScroll.js
@@ -1,0 +1,9 @@
+export default function enableScroll() {
+	if (window.removeEventListener) {
+		window.removeEventListener('DOMMouseScroll', preventDefault, false);
+	}
+	window.onmousewheel = document.onmousewheel = null;
+	window.onwheel = null;
+	window.ontouchmove = null;
+	document.onkeydown = null;
+};
