+++ templates/src/js/utils/preventDefault.js
@@ -1,0 +1,7 @@
+export default function preventDefault(e) {
+	e = e || window.event;
+	if (e.preventDefault) {
+		e.preventDefault();
+	}
+	e.returnValue = false;
+};
